+++ asyncgit/src/sync/mod.rs
@@ -1,6 +1,7 @@
 //! sync git api
 
 mod branch;
+mod commit;
 mod commit_details;
 mod commit_files;
 mod commits_info;
@@ -16,6 +17,7 @@ mod tags;
 pub mod utils;
 
 pub use branch::get_branch_name;
+pub use commit::{amend, get_head};
 pub use commit_details::{get_commit_details, CommitDetails};
 pub use commit_files::get_commit_files;
 pub use commits_info::{get_commits_info, CommitId, CommitInfo};
@@ -28,8 +30,8 @@ pub use reset::{reset_stage, reset_workdir};
 pub use stash::{get_stashes, stash_apply, stash_drop, stash_save};
 pub use tags::{get_tags, Tags};
 pub use utils::{
-    commit, is_bare_repo, is_repo, stage_add_all, stage_add_file,
-    stage_addremoved,
+    commit, commit_new, is_bare_repo, is_repo, stage_add_all,
+    stage_add_file, stage_addremoved,
 };
 
 #[cfg(test)]